Unify approach on verifying unsaved configs
We have produced a lot of screens which were using pretty much same
"safe_exit" code, which can be "streamlined" by creating a parent class
in common.

This will save us from need to re-implement the same functionality
on each screen independently.

We also solve error handling issue in networks,
where save can be reported as successfull when it was not made.

Signed-off-by: Dmitriy Rabotyagov <dmitriy@adria-cloud.com>